+++ src/services/auth-guard.service.ts
@@ -12,7 +12,7 @@ export class AuthGuardService {
     private router: Router
 
   ) { }
-  // checks for bearer token when navigating to pages - if no bearer token you will be redirected to login
+  // checks for bearer token when navigating to pages - if no bearer token you will be redirected to login page
   canActivate(route: ActivatedRouteSnapshot) {
     const bearer_token = localStorage.getItem("bearer_token");
     if (!bearer_token) {